+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-08T06:16:46.713Z",
+    "2026-07-08T07:52:33.694Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {