+++ packages/orq-rc/RELEASES.md
@@ -10619,3 +10619,13 @@ Based on:
 - [typescript v4.12.0-rc.6] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.6 - packages/orq-rc
+
+## 2026-07-08 07:55:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.7] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.7 - packages/orq-rc